Initial commit
I used a plant watering app over the past half year, and it was terribly
buggy. A few weeks ago they released an update, which made things a lot
worse, to the point where the app was not really usable any more.

I switched to a different app, which initially looked promising,
but it turned out to be so buggy that it was failing at its sole
responsibility. For some plants the watering reminder would just
disappear; deleting and re-adding the plant would not fix it.
Notifications did not fire, the calendar view was inconsistent with
the plant view; looking ahead for the upcoming week displayed nothing,
even though there were definitely plants to water. It was shameful.

Frustrated and disappointed with the state of plant watering apps,
I am thinking "how hard can this be". Keep the next watering date,
and when you tick it off, increment. I'll put it in SQLite and add
a basic server-side rendered webinterface, because last time I looked
into Android, I spent two days trying to set up the build toolchain
in a reproducible manner, and failed.

When you want something done well, you need to do it yourself.
So here we are.