Add module to read species from toml file
I want some format to define the species, and their properties, like
when they need watering. Committing a SQLite database to the repository
is not very friendly for diffs and merging, it needs to be a text-based
format. Json would be nice, but I want to be able to add comments. Yaml
has so many pitfalls that it is best to avoid it entirely. Toml is more
sane, let's try this.

Writing a parser with "tomland" was nice, just a few lines of code, easy
to get started, and it worked at the first try. Next step: inserting
these into the database so I can join the info with the plants.