Insert all species into database at startup
I thought at first that having them in the database makes it easy to
integrate the watering schedule into queries. But I am wondering now
whether I shouldn't just compute everything in Haskell, and do the join
manually with a hash table lookup. That will save me a lot of data
wrangling, getting the rows back out and adding them to the plant type.
Maybe I'll do that after all. But for now, here they are in the DB.