Load entire database into memory
At first I thought I could write queries to get things like the last
water event, event count, first event, etc, per plant. But then I
realized, to keep the list stable while still rendering html server-side
and doing a POST when watering, I want to ignore recent events for the
ordering (so watered plants don't immediately jump to the end of the
list). So the queries get more and more complicated. It can be done for
sure, but it is easier to just load everything into memory in Haskell
and there I have all flexibility I need. I suspectd data volume would
not be an issue even if I had hundreds of plants with years of watering
history. And when it does become a problem, it is easy to optimize.