Add ideas for future work
It would be nice to make it client side, because then I can just host a
static html page, and that would make the app usable for people who are
unable to host and secure their own server. By making everything
client-side, I don't have to deal with managing user accounts, which I
definitely do not want to do.

Having a client-side app will also simplify computations with local
dates, because you know what the local time zone is.

But having the data only client-side in local storage can be fragile
too, you do want to be able to make backups. We can add a server for
that. It would also enable syncing. The server does not need to know the
client data, it can be encrypted client-side, the server only needs to
act as a key-value store.

The next observation is that the data model is already append-only, so
it naturally forms a CRDT. (I would need to replace the sequential plant
ids with uuids, but I can do that at upgrade time.) That makes syncing
work even when there are multiple devices with changes involved. Perhaps
it is overkill but it is very elegant, and a nice excuse for me to
experiment with this.

Finally, once you have sync, you might want to sync with multiple people
who take care of the same plants. In that case it is good to add a
username and record who watered the plant. The username would only end
up in the local data, encrypted before the server sees it.