Add skeleton of PureScript app
A server-side app works for now, but for this to be useful to others
than me (and also for my own convenience -- I don't want to deploy this
to a public webserver), it would be good if the app was all client-side.

Add an initial empty PureScript setup. This is mostly copied from
Mindec. The html-dom code is taken verbatim from Mindec (also authored
by me, licensed Apache 2.0). I had already discovered there that Halogen
and animation don't go so well together, so I'll just do the dom
manipulation manually this time too.