Change to different getJulianDay implementation
I suspect mine is broken because division is not the right kind of
division. And indeed, negative numerators do occur. So I switched to a
different formula, which looks a bit simpler, but it still has a
negative numerator in one place, so it might still be wrong. In any
case, this does not solve my weird cactus bug.