Fix overflow bug in Time.addDays
An Int in PureScript is a 32-bit signed integer. And 2^31 milliseconds
is about 25 days. So when I needed to water my cactus in 30 days, that
wrapped around. The solution is to use the Number type, which can
represent integers up to 53 bits, which is sufficient for ranges of a
few hundred thousand years.

Would have been nice to have a debug mode with overflow checks, like
Rust has, to catch this issue early.

Also, the Gregorian day computation was not the problem after all. But
still the new version looks nicer, so I'll keep that one.