Add icons for watering buttons
Not sure if this is the way to go, I find it difficult to find a good
look for this app. The icons could also be solid, that might work
better. And I have a feeling I want something with green plant
photographs and white text, but I don't have the right content for it.

Either way, let's go with this for now.

The watering can icon is the iconic Ikea PS 2002.