Add routing to client-side app
I was playing with the idea of a single page app and nice app-like
transitions, but it breaks the back button, and it is a huge pain to get
it working in combination with scrolling with html/css, so for now I
will do the simple thing and have one detail page per plant. I can still
change it later if needed, but this is the fastest way to get something
to work.