Add an animation to status change
This way it does not "pop" so much, it is nice and smooth, and it also
acts a bit more as a visual indicator that something changed, because it
draws a bit of attention to it.

The info block still pops though. I need to restructure that one to
allow for a nicer transition, but the same transition is possible in
principle.