Begin building PureScript wrapper for IndexedDB
I want a fully local app, and apparently that involves service workers
nowadays, and service workers can only access IndexedDB, not local
storage, so I will have to use it for storage. It has a convoluted
interface with callbacks everywhere. Fortunately we can wrap that with
Aff, so the PureScript code will be nice and tidy.