Add build step for minified js
The output is not yet usable, the variable "PS" is no longer defined
after minifying. I suspect that Closure renames it to something shorter,
and we need to tell it to leave that particular name alone, but I
haven't yet figured out how. The minified file is about half the size
of the unminified file.