Add service worker, sort out minification in build
For the service worker, I needed js code to call into PureScript, so it
could register PureScript funtions as event handlers for the service
worker events. So I just concatenated the entry point with the compiled
and pruned PureScript code. That had the nice side effect that Google
Closure does not delete the entire thing as dead code, it can actually
minify and optimize things further. So let's hook that into the build
pipeline too now. The only downside is that for development, it is
rather slow. But I can work with the non-minified versions in
development.