Upgrade to latest commit in nixos-unstable channel
  atk              2.35.1   -> 2.36.0
  closure-compiler 20200315 -> 20200406
  cups             2.3.1    -> 2.3.3
  liberation-fonts 2.00.4   -> 2.1.0
  libglvnd         1.2.0    -> 1.3.1
  libjpeg-turbo    2.0.3    -> 2.0.4
  nettle           3.5.1    -> 3.6
  openssl          1.1.1f   -> 1.1.1g
  stack            2.1.3.1  -> 2.3.1
  systemd          243.7    -> 245.5
  util-linux       2.33.2   -> 2.35.1

Why does the development environment depend on systemd? I have no clue,
and I haven't discovered the incantation that makes "nix why-depends"
explain to me why. The liberation-fonts is probably due to being
uncareful in the image compressor. Cups may be related to that too.

There is a lot of stuff here that probably shouldn't be here, but it
works, and I'm not going to focus on reducing the closure size right now.