Add command for exporting species catalog as json
I mainly want to use this to run on CI, to check that all toml files can
be parsed. It is also useful to pre-generate the species file so it can
be served statically, which will simplify deploying the server, because
you don't need to take all of the species files and make them available
at runtime; the webserver can just serve it along with the js and html.