Add Epipremnum aureum
Wow, this is one of the messiest plants I have found so far. So much
inconsistent information on the interwebs. Is Epipremnum the same as the
Scindapsus? Is Marble Queen different from the Aureum? If yes, is it
true that they have very different watering needs? Does it like moist
soil, or do I have to let it dry out?

So take all of this with a grain of salt, we'll see how it goes.