Add a changelog
I want to continue working on local storage of the state, but with all
of the service worker stuff, I will break the current usage. And I still
want to be able to water my plants in the meantime, so I need to have a
stable point to go back to and use, while I continue to develop the next
version.

Also, having a version is useful for versioning the cache. Having
releases every once in a while acts as a compatibility reference point.

I don't expect anybody to use or package this, especially not at this
point, I am mainly adding this for myself.