Store watering events locally
It doesn't work as it should yet, the last modification is not saved,
although earlier ones are. I suspect I need to commit, and if I don't,
beginning a new transaction triggers committing the previous one?

In any case, this paves the way to storing everything fully locally, so
I no longer need to spin up a server to water my plants, and so multiple
people can use this app without me having to manage accounts.