Lower watering interval for basilicum
I watered mine too little and this morning all of the leaves were
crumbled and hanging. Now that I watered it, it is recovering. Might
have something to do with my pot as well, but 6 days was definitely too
low.

Also, I have an idea for adaptive watering now. If the reminder is
always too early, we can detect that from past waterings, and adapt the
watering scheme to change the interval.