Try to include more files in CI cache keys
So after bumping the Nixpkgs snapshot, the existing caches are no longer
useful (not harmful either, fortunately). We should include the pinned
Nixpkgs snapshot in the hash then, but is it possible to hash multiple
files with "hashFiles"? The documentation mentions globs, but there
doesn't appear to be a way to hash multiple files if you can't come up
with a glob that matches exactly those. Maybe just a space as separator
will work? The only way to know is to try.