Pass multiple arguments to hashFile on CI after all
I was looking at the wrong place in the logs; it prints "restored from
key", but that is not the original cache key it tried to push to, deeper
in the logs is the original key. With a space in the glob, the hash
became empty string, so I suppose that was wrong. But when passing two
arguments, the hash was different than the first time.

This reverts commit 19cda2ec2362968e1d92b471ef68610056ae22ff.