Drop Stack in favor of GHC with Nix
I was getting this "file not found" error for the binary in .stack-work,
while the file definitely existed. Not sure what was going on here, but
instead of spending the time to debug, let's just drop Stack entirely
and use raw GHC; it's not a complex Haskell project anyway, it's a
single file with few external dependencies, and Nix can provide those.