Do not respond to diagonal swipes
I added the "swipe to reveal the delete button", but it only looked at
the x-offset. Then when I started using it, ofen I would trigger this by
just scrolling the page, because your fingertip essentially moves on a
circle, so it is going to have a dx even when your intention is to only
have a dy.

Fix this by limiting the y-offset for which the swipe still triggers.