Use median for adaptive interval instead of average
I have the feeling that I'm more often late with watering than early,
and regularly this is by accidenty, not because the soil isn't dry yet.
Using the median should reduce the impact that such outliers have.
Especially if I'm away for a few days and the interval doubles, it
should not affect the median while it would affect the average much
more.